+++ index.js
@@ -1,380 +1,427 @@
 var Service, Characteristic;
-var mqtt = require("mqtt");
-const packageJson = require('./package.json')
+var mqtt = require('mqtt');
+const packageJson = require('./package.json');
 
 module.exports = function (homebridge) {
 	Service = homebridge.hap.Service;
 	Characteristic = homebridge.hap.Characteristic;
 
-	homebridge.registerAccessory("homebridge-homeqtt-alarm", "homebridge-homeqtt-alarm", homeqttAlarmAccessory);
+	homebridge.registerAccessory('homebridge-homeqtt-alarm', 'homebridge-homeqtt-alarm', homeqttAlarmAccessory);
+};
+
+// State and Command Payloads
+const stateStayArm = 'STAY_ARM'; // 0
+const stateAwayArm = 'AWAY_ARM'; // 1
+const stateNightArm = 'NIGHT_ARM'; // 2
+const stateDisarmed = 'DISARMED'; // 3
+const stateDisarm = 'DISARM'; // 3
+const stateTriggered = 'ALARM_TRIGGERED'; // 4
+
+const commandStayArm = 0; // STAY_ARM = 0
+const commandAwayArm = 1; // AWAY_ARM = 1
+const commandNightArm = 2; // NIGHT_ARM = 2
+const commandDisarmed = 3; // DISARMED = 3
+const commandTriggered = 4; // ALARM_TRIGGERED = 4
+
+function stateName(msg) {
+	switch (msg) {
+		case commandStayArm:
+			msg = stateStayArm;
+			break;
+		case commandAwayArm:
+			msg = stateAwayArm;
+			break;
+		case commandNightArm:
+			msg = stateNightArm;
+			break;
+		case commandDisarmed:
+			msg = stateDisarmed;
+			break;
+		case commandTriggered:
+			msg = stateTriggered;
+			break;
+		default:
+			msg = null;
+			break;
+	}
+	return msg;
+}
+
+function getAllowedTargetStates(targetStates) {
+	var allowedTargetStates = [3]; // Always have disarmed
+	for (var state in targetStates) {
+		if (targetStates[state]) {
+			if (state == 'stayArm') {
+				state = 0;
+				allowedTargetStates.push(state);
+			}
+			if (state == 'awayArm') {
+				state = 1;
+				allowedTargetStates.push(state);
+			}
+			if (state == 'nightArm') {
+				state = 2;
+				allowedTargetStates.push(state);
+			}
+		}
+	}
+	return allowedTargetStates;
 }
 
 function homeqttAlarmAccessory(log, config) {
-	// Default Variables
-	this.log = log
+	this.log = log;
 
 	// System Settings
-	this.manufacturer = '@nzbullet'
-	this.model = 'Homeqtt Alarm'
+	this.manufacturer = '@nzbullet';
+	this.model = 'Homeqtt Alarm';
 
-	// State and Command Payloads
-	this.stateStayArm = 'STAY_ARM'; // 0
-	this.stateAwayArm = 'AWAY_ARM'; // 1
-	this.stateNightArm = 'NIGHT_ARM'; // 2
-	this.stateDisarmed = 'DISARMED'; // 3
-	this.stateDisarm = 'DISARM'; // 3
-	this.stateTriggered = 'ALARM_TRIGGERED'; // 4
-
-	this.commandStayArm = 0 // 'STAY_ARM';
-	this.commandAwayArm = 1 // 'AWAY_ARM';
-	this.commandNightArm = 2 //' NIGHT_ARM';
-	this.commandDisarmed = 3 // 'DISARMED';
-	this.commandTriggered = 4 // 'ALARM_TRIGGERED';
-
-	// Load data from config.json
-	// Name of sensor in HomeKit (Defaults to 'Homeqtt')
-	if (!config.name) {
-		this.name = 'Homeqtt'
-		log('Setup: Name not provided. Default name being used - Homeqtt')
-	}
-	// Homebridge Accessory
-	if (!config.accessory) {
-		config.accessory = 'homebridge-homeqtt-alarm'
+	// No Config Exit
+	if (!config) {
+		return;
 	}
 	// Debug (Extra logging)
-	if (config.debug) {
-		this.debug = config.debug;
+	if (config.debug === true) {
+		this.debug = log
 	} else {
-		config.debug = false;
-		this.debug = config.debug;
+		this.debug = function(){}
 	}
+
+	// Name of sensor in HomeKit (Defaults to 'Homeqtt')
+	if (!config.name) {
+		this.name = 'Homeqtt';
+		this.debug('Setup: Name not provided. Default name being used - Homeqtt');
+	}
+
+	// Homebridge Accessory
+	if (!config.accessory) {
+		config.accessory = 'homebridge-homeqtt-alarm';
+	}
+
 	// MQTT Config
 	if (config.mqttConfig) {
-		this.mqttConfig = config.mqttConfig
-		if (!config.mqttConfig.url) {
-			log('Stopping: You have not set your MQTT Server URL in your config.json')
+		var url = config.mqttConfig.url;
+		var username = config.mqttConfig.username;
+		var password = config.mqttConfig.password;
+		if (!url) {
+			log('Stopping: You have not set your MQTT Server URL in your config.json');
 			return; // Error
-		}
-	} else {
-		log('Stopping: You have not setup your MQTT Config in your config.json')
-		return; // Error
-	}
-	// MQTT Broker connection settings
-	if (this.mqttConfig && config.mqttConfig.url) {
-		var mqttClientId = config.name.replace(/[^\x20-\x7F]/g, "") + '_' + Math.random().toString(16).substr(2, 8);
-		var mqttBroker = config.mqttConfig.url // MQTT Broker URL and Port
-		var options = {
-			keepalive: 45,
-			clientId: mqttClientId,
-			protocolId: 'MQTT',
-			protocolVersion: 4,
-			clean: true,
-			reconnectPeriod: 1000,
-			connectTimeout: 30 * 1000,
-			will: {
-				topic: 'WillMsg',
-				payload: 'Connection closed abnormally!',
-				qos: 0,
-				retain: false
-			},
-			username: config.mqttConfig.username,
-			password: config.mqttConfig.password,
-			rejectUnauthorized: false
-		};
-		// Show MQTT Options in log
-		if (this.debug) {
-			log('Setup: MQTT Broker: ' + mqttBroker);
-			log('Setup: MQTT Options: ' + JSON.stringify(options, function (key, value) {
-				if (key == "password") {
-					return undefined; // filter out
-				}
-				return value;
-			}));
-		}
-		// Connect to MQTT 
-		try {
-			this.client = mqtt.connect(mqttBroker, options);
-		} catch (e) {
-			log.error(e);
-			return;
-		}
-		// MQTT Connection Error
-		this.client.on('error', function () {
-			log('Error event on MQTT');
-		});
-	} else {
-		log('Stopping: MQTT Has not been set up in config.json')
-		return; // Error
-	}
-	// MQTT Topics
-	if (config.alarmTopics) {
-		// Settings Topics (does not include RF Key Topics)
-		this.alarmTopics = config.alarmTopics
-		// Message Topic
-		if (!config.alarmTopics.messageTopic) {
-			config.alarmTopics.messageTopic = 'tele/rfbridge/RESULT'
-			this.messageTopic = config.alarmTopics.messageTopic
-			log('Setup: Message topic not provided. Default being used: tele/rfbridge/RESULT')
-		}
-		// Target State Topic
-		if (!config.alarmTopics.setTargetStateTopic) {
-			config.alarmTopics.setTargetStateTopic = 'alarm/target'
-			this.setTargetStateTopic = config.alarmTopics.setTargetStateTopic
-			log('Setup: Target State topic not provided. Default being used: alarm/target')
-		}
-		// Current State Topic
-		if (!config.alarmTopics.getCurrentStateTopic) {
-			config.alarmTopics.getCurrentStateTopic = 'alarm/current'
-			this.getCurrentStateTopic = config.alarmTopics.getCurrentStateTopic
-			log('Setup: Current State topic not provided. Default being used: alarm/current')
-		}
-	} else {
-		log('Stopping: You have not setup your Alarm Topics in your config.json')
-		return; // Error
-	}
-	// Alarm Settings
-	if (config.alarmSettings) {
-		this.alarmSettings = config.alarmSettings
-		if (config.alarmSettings.targetStates) {
-			// Target States
-			this.targetStates = config.alarmSettings.targetStates
-			if (!config.alarmSettings.targetStates.stayArm) {
-				this.alarmSettings.targetStates.stayArm = false
-			}
-			if (!config.alarmSettings.targetStates.awayArm) {
-				this.alarmSettings.targetStates.awayArm = false
-			}
-			if (!config.alarmSettings.targetStates.nightArm) {
-				this.alarmSettings.targetStates.nightyArm = false
-			}
 		} else {
-			log('Stopping: You have not setup your Target States under Alarm Settings in your config.json')
-			return; // Error
-		}
-	} else {
-		log('Stopping: You have not setup your Alarm Settings in your config.json')
-		return; // Error
-	}
-	// Keyfob Enabled / Disabled
-	if (config.sensor) {
-		this.sensor = config.sensor // All Sensors Object
-	} else {
-		log('Stopping: You have not setup your Sensors in your config.json')
-		return; // Error
-	}
-	if (config.keyfob) {
-		this.keyfob = config.keyfob
-	} else {
-		this.keyfob = false
-	}
-	if (this.keyfob) {
-		if (config.keyfobs) {
-			this.keyfobs = config.keyfobs // Keyfob object
-			if (config.keyfobs.buttons) {
-				this.buttons = config.keyfobs.buttons // Button object
+			var mqttClientId = config.name.replace(/[^\x20-\x7F]/g, "") + '_' + Math.random().toString(16).substr(2, 8);
+			var mqttBroker = url; // MQTT Broker URL and Port
+			var options = {
+				keepalive: 45,
+				clientId: mqttClientId,
+				protocolId: 'MQTT',
+				protocolVersion: 4,
+				clean: true,
+				reconnectPeriod: 1000,
+				connectTimeout: 30 * 1000,
+				will: {
+					topic: 'WillMsg',
+					payload: 'Connection closed abnormally!',
+					qos: 0,
+					retain: false
+				},
+				username: username,
+				password: password,
+				rejectUnauthorized: false
+			};
+			// Connect to MQTT 
+			try {
+				this.client = mqtt.connect(mqttBroker, options);
+			} catch (err) {
+				log(err); // Not shown in the log ever
+				return;
 			}
 		}
-	}
-	if (config.siren) {
-		this.siren = config.siren // Siren object
 	} else {
-		this.siren.enabled = false
+		log('Stopping: You have not setup your MQTT Config in your config.json');
+		return; // Error
 	}
-	// Keyfob
-	if (this.keyfob === true) {
-		log("Setup: Keyfob available")
-		// Siren
-		if (this.siren.enabled === true) {
-			log("Setup: Siren Enabled")
-		} else {
-			log("Setup: Siren Disabled")
-		};
-	} else {
-		log("Setup: No Keyfob available")
-	};
-	
-	// Set initial state to disarmed
-	log('Setup: Setting initial HomeKit state to disarmed');
-	this.readstate = Characteristic.SecuritySystemCurrentState.DISARMED;
 
 	// Set 'that' to 'this' to be able to access global 'this' inside 'this'
-	var that = this
+	var that = this;
 
+	// MQTT Errors
+	this.client.stream.on('error', (err) => {
+		if (config.debug === true) {
+			that.debug('MQTT', err);
+		} else {
+			log('MQTT Error. Please check you config.json or broker. For more detail enable debug.');
+		}
+		this.client.end();
+		return;
+	});
+	this.client.on('error', (err) => {
+		log('Error event on MQTT: ', err);
+		this.client.end();
+	});
+
+	// MQTT Topics (does not include RF Key Topics)
+	if (config.alarmTopics) {
+		this.alarmTopics = config.alarmTopics; // All Topics
+		this.messageTopic = config.alarmTopics.messageTopic; // Message Topic
+		this.setTargetStateTopic = config.alarmTopics.setTargetStateTopic; // Target State Topic
+		this.getCurrentStateTopic = config.alarmTopics.getCurrentStateTopic; // Current State Topic
+		// Set Default Topics if missing
+		if (!this.messageTopic) { // Message Topic Blank
+			config.alarmTopics.messageTopic = 'tele/rfbridge/RESULT';
+			this.messageTopic = config.alarmTopics.messageTopic;
+			that.debug('Setup: Message topic not provided. Default being used: tele/rfbridge/RESULT');
+		}
+		if (!this.setTargetStateTopic) { // Target State Topic
+			config.alarmTopics.setTargetStateTopic = 'alarm/target';
+			this.setTargetStateTopic = config.alarmTopics.setTargetStateTopic;
+			that.debug('Setup: Target State topic not provided. Default being used: alarm/target');
+		}
+		if (!this.getCurrentStateTopic) { // Current State Topic
+			config.alarmTopics.getCurrentStateTopic = 'alarm/current';
+			this.getCurrentStateTopic = config.alarmTopics.getCurrentStateTopic;
+			that.debug('Setup: Current State topic not provided. Default being used: alarm/current');
+		}
+	} else {
+		log('Stopping: You have not setup your Alarm Topics in your config.json');
+		return; // Error
+	}
+
+	// Alarm Settings
+	if (config.alarmSettings) {
+		this.alarmSettings = config.alarmSettings;
+		if (this.alarmSettings.targetStates) {
+			this.targetStates = config.alarmSettings.targetStates;
+		} else {
+			log('Stopping: You have not setup targetStates in your config.json');
+			return false;
+		}
+	} else {
+		log('Stopping: You have not setup alarmSettings in your config.json');
+		return; // Error
+	}
+	// Sensors
+	if (config.sensor) {
+		this.sensor = config.sensor; // All Sensors Object
+	} else {
+		log('Stopping: You have not setup your Sensors in your config.json');
+		return; // Error
+	}
+	// Keyfob
+	if (config.keyfob) {
+		this.keyfob = config.keyfob;
+		if (config.keyfobs) {
+			this.keyfobs = config.keyfobs;
+			// Buttons
+			if (config.keyfobs.buttons) {
+				this.buttons = config.keyfobs.buttons;
+			}
+		}
+	} else {
+		this.keyfob = false;
+	}
+	// Siren
+	if (config.siren) {
+		this.siren = config.siren;
+	} else {
+		this.siren.enabled = false;
+	}
+
+	// Connect and Subscribe to all topics on connect
+	this.client.on('connect', function () {
+		// Show MQTT Options in log
+		that.debug('Setup: MQTT Broker: ' + mqttBroker);
+		if (that.keyfob) {
+			that.debug('Setup: Keyfob Available');
+		}
+		if (that.siren.enabled) {
+			that.debug('Setup: Siren Enabled');
+		} else {
+			that.debug('Setup: Siren Disabled');
+		}
+		for (let topic in that.alarmTopics) {
+			if (topic) {
+				that.client.subscribe(that.alarmTopics[topic]);
+				that.debug('Setup: Connected and subscribed to:', that.alarmTopics[topic]);
+			}
+		}
+		if (that.keyfob) {
+			for (let keyfob in that.keyfobs) {
+				var buttons = that.keyfobs[keyfob].buttons;
+				if (that.keyfobs[keyfob].enabled) {
+					for (let button in buttons) {
+						button = buttons[button];
+						if (button.enabled) {
+							let allowedTargetStates = getAllowedTargetStates(that.targetStates);
+							if (allowedTargetStates.includes(button.alarmState) || button.alarmState === 4) {
+								let state = stateName(button.alarmState);
+								if (button.rfkeyTopic) {
+									that.client.subscribe(button.rfkeyTopic);
+									that.debug('Setup: Connected and subscribed to RFKey topic:', button.rfkeyTopic);
+								} else {
+									log('Warning:', state, 'Button RFKey Topic has not been provided in config.json for', keyfob.name);
+									return;
+								}
+							}
+						}
+					}
+				}
+			}
+		}
+		// Enable States
+		var targetstates = Object.keys(that.targetStates).filter(key => that.targetStates[key])
+		that.debug('Setup: Enabled Alarm States -',targetstates);
+		// Set initial state to disarmed
+		this.readstate = Characteristic.SecuritySystemCurrentState.DISARMED;
+		that.debug('Setup: Setting initial HomeKit state to DISARMED (', this.readstate, ')');
+		log('Setup Complete');
+	});
 	// MQTT Message Received
 	this.client.on('message', function (topic, msg) {
 		var message = msg.toString();
 		// Set HomeKit Alarm based on Keyfob Button presses
-		if (that.keyfob === true && topic === that.alarmTopics.messageTopic) {
+		if (that.keyfob && topic === that.messageTopic) {
 			for (let keyfob in that.keyfobs) {
-				if (that.keyfobs[keyfob].enabled === true) {
-					for (let button in that.keyfobs[keyfob].buttons) {
-						if (that.keyfobs[keyfob].buttons[button].enabled === true && message.indexOf(that.keyfobs[keyfob].buttons[button].MQTTCode) !== -1) {
-							var stateName = that.keyfobs[keyfob].buttons[button].alarmState
-							switch (stateName) {
-								case that.commandStayArm:
-									stateName = 'STAY_ARM'
-									break;
-								case that.commandAwayArm:
-									stateName = 'AWAY_ARM'
-									break;
-								case that.commandNightArm:
-									stateName = 'NIGHT_ARM'
-									break;
-								case that.commandDisarmed:
-									stateName = 'DISARMED'
-									break;
-								case that.commandTriggered:
-									stateName = 'ALARM_TRIGGERED'
-									break;
-								default:
-									stateName = null;
-									break;
+				var buttons = that.keyfobs[keyfob].buttons;
+				if (that.keyfobs[keyfob].enabled) {
+					for (let button in buttons) {
+						button = buttons[button];
+						if (button.enabled && message.indexOf(button.MQTTCode) !== -1) {
+							let state = stateName(button.alarmState);
+							that.debug('Keyfob Button Pressed:', state, '(', button.alarmState, ')');
+							if (state === 'ALARM_TRIGGERED') {
+								that.securityService.setCharacteristic(Characteristic.SecuritySystemCurrentState, commandTriggered);
+								that.client.publish(that.getCurrentStateTopic, stateTriggered);
+								if (that.keyfob && that.siren.enabled) {
+									log('Siren Status: Triggered');
+								}
+								log('SOS Triggered: Keyfob SOS Button Pressed');
 							}
-							if (that.debug) {
-								log("Keyfob Button Pressed:", stateName, "(", that.keyfobs[keyfob].buttons[button].alarmState, ")")
-							}
-							if (stateName === 'ALARM_TRIGGERED') {
-								that.securityService.setCharacteristic(Characteristic.SecuritySystemCurrentState, that.commandTriggered);
-								that.client.publish(that.alarmTopics.getCurrentStateTopic, that.stateTriggered);
-							}
-							if (stateName !== 'ALARM_TRIGGERED') {
-								that.securityService.setCharacteristic(Characteristic.SecuritySystemTargetState, that.keyfobs[keyfob].buttons[button].alarmState);
+							if (state !== 'ALARM_TRIGGERED') {
+								let allowedTargetStates = getAllowedTargetStates(that.targetStates);
+								if (allowedTargetStates.includes(button.alarmState)) {
+									that.securityService.setCharacteristic(Characteristic.SecuritySystemTargetState, button.alarmState);
+								} else {
+									that.debug('Button pressed:', state, '- Alarm state not enabled in HomeKit');
+								}
 							}
 						}
 					}
-
 				}
 			}
 		}
 		// Current state or target state topic messages (setting the state)
-		if (topic === that.alarmTopics.getCurrentStateTopic || topic === that.alarmTopics.setTargetStateTopic) {
-			if (that.debug) {
-				log("Topic:", topic)
-				log("Message:", message)
-			}
+		if (topic === that.getCurrentStateTopic || topic === that.setTargetStateTopic) {
+			that.debug('Topic:', topic);
+			log('Alarm State:', message);
 			switch (message) {
-				case that.stateStayArm:
-					message = Characteristic.SecuritySystemCurrentState.STAY_ARM; // STAY_ARM = 0
+				case stateStayArm:
+					message = commandStayArm;
 					break;
-				case that.stateAwayArm:
-					message = Characteristic.SecuritySystemCurrentState.AWAY_ARM; // AWAY_ARM = 1
+				case stateAwayArm:
+					message = commandAwayArm;
 					break;
-				case that.stateNightArm:
-					message = Characteristic.SecuritySystemCurrentState.NIGHT_ARM; // NIGHT_ARM = 2
+				case stateNightArm:
+					message = commandNightArm;
 					break;
-				case that.stateDisarmed:
-					message = Characteristic.SecuritySystemCurrentState.DISARMED; // DISARMED = 3
+				case stateDisarmed:
+					message = commandDisarmed;
 					break;
-				case that.stateDisarm:
-					message = Characteristic.SecuritySystemCurrentState.DISARMED; // DISARMED = 3
+				case stateDisarm:
+					message = commandDisarmed;
 					break;
-				case that.stateTriggered:
-					message = Characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED; // ALARM_TRIGGERED = 4
+				case stateTriggered:
+					message = commandTriggered;
 					break;
 				default:
 					message = null;
 					break;
-			};
+			}
 			// Set HomeKit state to alarm state
 			if (message !== null) {
 				that.readstate = message;
 				that.securityService.getCharacteristic(Characteristic.SecuritySystemCurrentState, that.readstate);
 				// Set Siren State if siren enabled
-				if (that.keyfob === true && that.siren.enabled === true) {
-					for (let keyfob in that.keyfobs) {
-						if (that.keyfobs[keyfob].enabled === true) {
-							for (let button in that.keyfobs[keyfob].buttons) {
-								if (that.keyfobs[keyfob].buttons[button].enabled === true && that.keyfobs[keyfob].buttons[button].alarmState === that.readstate) {
-									// On change of HomeKit switch publish to all RFKey topics for that state (for enabled buttons)
-									// This is because HomeKit doesn't know which RFKey or MQTT Code to use 
-									that.client.publish(that.keyfobs[keyfob].buttons[button].rfkeyTopic, that.keyfobs[keyfob].buttons[button].MQTTCode);
-									if (that.debug) {
-										log("MQTT Code", that.keyfobs[keyfob].buttons[button].MQTTCode, "published to", that.keyfobs[keyfob].buttons[button].rfkeyTopic)
-									}
-								}
-							}
-						}
-					}
-					if (that.debug) {
-						if (that.readstate === 0 || that.readstate === 1 || that.readstate === 2) { // Armed
-							log("Siren Status: Armed")
-						}
-						if (that.readstate === 3) { // Disarmed
-							log("Siren Status: Disarmed")
-						}
-						if (that.readstate === 4) { // SOS
-							log("Siren Status: Triggered")
-						}
-					}
-				}
-			};
-		}
-		// Sensor topic messages (triggered messages)
-		if (topic == that.alarmTopics.messageTopic) {
-			for (let sensor in that.sensor) {
-				// Sensor is in MQTT message
-
-				if (message.indexOf(that.sensor[sensor].MQTTCode) !== -1) {
-					// Sensor is enabled and alarm is not Disarmed
-					if (that.sensor[sensor].enabled === true && that.readstate != 3) {
-						// If sensor is allowed in the enabled state (and alarm is triggered)
-						if (that.sensor[sensor].allowStay === true && that.readstate === 0) {
-							triggerAlarm.call()
-						}
-						if (that.sensor[sensor].allowAway === true && that.readstate === 1) {
-							triggerAlarm.call()
-						}
-						if (that.sensor[sensor].allowNight === true && that.readstate === 2) {
-							triggerAlarm.call()
-						}
-						// Trigger Alarm Function
-						function triggerAlarm() {
-							log('Sensor Triggered:', that.sensor[sensor].location, '(', that.sensor[sensor].MQTTCode, ')')
-							// MQTT Publish Triggered 
-							that.client.publish(that.alarmTopics.getTargetStateTopic, that.stateTriggered);
-							// Trigger Alarm in HomeKit
-							that.securityService.setCharacteristic(Characteristic.SecuritySystemCurrentState, that.commandTriggered);
-							// Trigger Siren if enabled
-							if (that.keyfob === true && that.siren.enabled === true) {
-								for (let keyfob in that.keyfobs) {
-									if (that.keyfobs[keyfob].enabled === true) {
-										for (let button in that.keyfobs[keyfob].buttons) {
-											if (that.keyfobs[keyfob].buttons[button].enabled === true && that.keyfobs[keyfob].buttons[button].alarmState === that.commandTriggered) {
-												that.client.publish(that.keyfobs[keyfob].buttons[button].rfkeyTopic, that.keyfobs[keyfob].buttons[button].MQTTCode);
+				var set = 0;
+				while (set < 1) {
+					if (that.keyfob && that.siren.enabled) {
+						for (let keyfob in that.keyfobs) {
+							if (keyfob) {
+								if (that.keyfobs[keyfob].enabled) {
+									var buttons = that.keyfobs[keyfob].buttons;
+									for (let button in buttons) {
+										button = buttons[button]
+										if (button.enabled && button.alarmState === that.readstate) {
+											let state = stateName(button.alarmState);
+											let allowedTargetStates = getAllowedTargetStates(that.targetStates);
+											if (allowedTargetStates.includes(button.alarmState) || that.readstate === 4) {
+												that.client.publish(button.rfkeyTopic, button.MQTTCode);
+												set++;
+												that.debug('MQTT Code', button.MQTTCode, 'published to', button.rfkeyTopic);
+											} else {
+												that.debug('Button pressed:', state, '- Alarm state not enabled in HomeKit. MQTT Code not sent.');
 											}
 										}
 									}
 								}
-								log("Siren Triggered")
+								if (set === 1) {
+									break;
+								}
 							}
 						}
+					}
+					break;
+				}
+				if (config.debug === true) {
+					let allowedTargetStates = getAllowedTargetStates(that.targetStates);
+					if (allowedTargetStates.includes(that.readstate) || that.readstate === 4) {
+						let sirenState = stateName(that.readstate);
+						that.debug('Siren Status:', sirenState);
 					}
 				}
 			}
 		}
-	});
-
-	// Subscribe to all topics on connect
-	this.client.on('connect', function () {
-		for (let topic in that.alarmTopics) {
-			that.client.subscribe(that.alarmTopics[topic]);
-			if (that.debug) {
-				log('Setup: Connected and subscribed to: ', that.alarmTopics[topic]);
+		// Trigger Alarm Function
+		function triggerAlarm() {
+			// MQTT Publish Triggered 
+			that.client.publish(that.setTargetStateTopic, stateTriggered);
+			// Trigger Alarm in HomeKit
+			that.securityService.setCharacteristic(Characteristic.SecuritySystemCurrentState, commandTriggered);
+			// Trigger Siren if enabled
+			var trigger = 0;
+			while (trigger < 1) {
+				if (that.keyfob && that.siren.enabled) {
+					for (let keyfob in that.keyfobs) {
+						if (keyfob) {
+							if (that.keyfobs[keyfob].enabled) {
+								var buttons = that.keyfobs[keyfob].buttons;
+								for (let button in buttons) {
+									button = buttons[button];
+									if (button.enabled && button.alarmState === commandTriggered) {
+										that.client.publish(button.rfkeyTopic, button.MQTTCode);
+										log('Siren Triggered');
+										trigger++;
+									}
+								}
+							}
+							if (trigger === 1) {
+								break;
+							}
+						}
+					}
+				}
+				break;
 			}
 		}
-		if (that.keyfob === true) {
-			for (let keyfob in that.keyfobs) {
-				if (that.keyfobs[keyfob].enabled === true) {
-					for (let button in that.keyfobs[keyfob].buttons) {
-						if (that.keyfobs[keyfob].buttons[button].enabled === true) {
-							if (that.keyfobs[keyfob].buttons[button].rfkeyTopic) {
-								that.client.subscribe(that.keyfobs[keyfob].buttons[button].rfkeyTopic);
-								if (that.debug) {
-									log('Setup: Connected and subscribed to RFKey topic: ', that.keyfobs[keyfob].buttons[button].rfkeyTopic);
-								}
-							} else {
-								log('Stopping: Button RFKey Topic has not been provided in config.json')
-							}
+		// Sensor topic messages (triggered messages)
+		if (topic == that.messageTopic) {
+			for (let sensor in that.sensor) {
+				sensor = that.sensor[sensor]
+				// Sensor is in MQTT message
+				if (message.indexOf(sensor.MQTTCode) !== -1) {
+					// Sensor is enabled and alarm is not Disarmed
+					if (sensor.enabled && that.readstate != 3) {
+						// If sensor is allowed in the enabled state (and alarm is triggered)
+						if (sensor.allowStay && that.readstate == 0 || sensor.allowAway && that.readstate == 1 || sensor.allowNight && that.readstate == 2) {
+							triggerAlarm();
+							log('Sensor Triggered:', sensor.location, '(', sensor.MQTTCode, ')');
 						}
 					}
 				}
@@ -386,61 +433,45 @@ function homeqttAlarmAccessory(log, config) {
 homeqttAlarmAccessory.prototype = {
 	// Get and Set states
 	setTargetState: function (state, callback) {
-		switch (state) {
-			case Characteristic.SecuritySystemTargetState.STAY_ARM:
-				mqttstate = this.stateStayArm;
-				break;
-			case Characteristic.SecuritySystemTargetState.AWAY_ARM:
-				mqttstate = this.stateAwayArm;
-				break;
-			case Characteristic.SecuritySystemTargetState.NIGHT_ARM:
-				mqttstate = this.stateNightArm;
-				break;
-			case Characteristic.SecuritySystemTargetState.DISARMED:
-				mqttstate = this.stateDisarmed;
-				break;
-			case Characteristic.SecuritySystemTargetState.DISARM:
-				mqttstate = this.stateDisarm;
-				break;
-		};
+		let mqttstate = stateName(state);
 		// MQTT Publish State 
-		this.client.publish(this.alarmTopics.getCurrentStateTopic, mqttstate, {
+		this.client.publish(this.getCurrentStateTopic, mqttstate, {
 			qos: 0,
 			retain: true
 		});
-		this.log("Setting state to:", mqttstate)
+		this.debug('Setting state to:', mqttstate, '(', state, ')');
 		this.securityService.setCharacteristic(Characteristic.SecuritySystemCurrentState, state);
 		callback(null, state);
 	},
 
-	getState: function (callback) {
-		if (this.readstate == "0" || this.readstate == "1" || this.readstate == "2" || this.readstate == "3" || this.readstate == "4") {
-			if (this.debug) {
-				this.log("Setting state to:", this.readstate);
+	getState: function (get, callback) {
+		let state = stateName(this.readstate);
+		if (this.readstate == '0' || this.readstate == '1' || this.readstate == '2' || this.readstate == '3' || this.readstate == '4') {
+			if (get == 'current') {
+				this.debug('Current Alarm State:', state, '(', this.readstate, ')');
 			}
+			if (get == 'target') {
+				this.debug('Target Alarm State:', state, '(', this.readstate, ')');
+			} 
 			callback(null, this.readstate);
 		} else {
-			this.log(this.readstate, " is not a valid HomeKit State. It should be one of: 0, 1, 2, 3, 4");
-			callback("error");
-		};
+			this.debug('Received an invalid HomeKit State:', this.readstate, '- Should be one of 0,1,2,3,4. Please check your config.json for errors.');
+			callback('error');
+		}
 	},
 
 	getCurrentState: function (callback) {
-		if (this.debug) {
-			this.log("Getting current state")
-		}
-		this.getState(callback);
+		let get = 'current'
+		this.getState(get, callback);
 	},
 
 	getTargetState: function (callback) {
-		if (this.debug) {
-			this.log("Getting target state")
-		}
-		this.getState(callback);
+		let get = 'target'
+		this.getState(get, callback);
 	},
 
 	identify: function (callback) {
-		this.log("Identify requested!");
+		this.log('Identify requested!');
 		callback(); // success
 	},
 
@@ -456,25 +487,9 @@ homeqttAlarmAccessory.prototype = {
 			.getCharacteristic(Characteristic.SecuritySystemCurrentState)
 			.on('get', this.getCurrentState.bind(this));
 		// Set allowed target states
-		var allowedTargetStates = [3] // Always have disarmed
-		for (var state in this.targetStates) {
-			if (this.targetStates[state] === true) {
-				if (state == "stayArm") {
-					state = 0
-					allowedTargetStates.push(state)
-				}
-				if (state == "awayArm") {
-					state = 1
-					allowedTargetStates.push(state)
-				}
-				if (state == "nightArm") {
-					state = 2
-					allowedTargetStates.push(state)
-				}
-				let characteristic = this.securityService.getCharacteristic(Characteristic.SecuritySystemTargetState);
-				characteristic.props.validValues = allowedTargetStates; // (needs to be a number array e.g. [0,1,2,3])
-			}
-		}
+		let allowedTargetStates = getAllowedTargetStates(this.targetStates);
+		let characteristic = this.securityService.getCharacteristic(Characteristic.SecuritySystemTargetState);
+		characteristic.props.validValues = allowedTargetStates; // (needs to be a number array e.g. [0,1,2,3])
 		this.securityService
 			.getCharacteristic(Characteristic.SecuritySystemTargetState)
 			.on('get', this.getTargetState.bind(this))